ui: Sidebar Menu -> add popover for create new pack
