+++ src/index.css
@@ -12,9 +12,11 @@
 	-moz-osx-font-smoothing: grayscale; */
 	--spinner-outer-bg-color: #f3f3f3;
 	--spinner-inner-bg-color: #949494;
-	--hover-dark: brightness(80%);
+	--hover-dark-2: brightness(80%);
 	--hover-dark-1: brightness(95%);
-	--hover-light: brightness(120%);
+	--hover-light-1: brightness(105%);
+	--hover-light-2: brightness(120%);
+	--hover-light-3: brightness(150%);
 	--max-width: 1280px;
 }
